fastapi-namaste-icd11/requirements.txt
fastapi>=0.111,<0.115
uvicorn[standard]>=0.29,<0.32
pydantic>=2.7,<3
python-jose>=3.3,<4
passlib[bcrypt]>=1.7,<2
bcrypt>=3.1.0,<4
pytest>=8,<9
python-multipart>=0.0.6,<0.1
httpx>=0.27,<0.28
